FIB_SEB_reader/fib_sem_measurement_tool/requirements.txt
customtkinter>=5.2.2
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

